Use SVG node attributes to wait for style change
Since @testing-library/jest-dom's toHaveStyle matcher doesn't work on
this SVG node for some reason, use the attributes of the node to check
for the style and wait for the fill to be green. Use of attributes
inspired by
https://github.com/facebook/jest/blob/d1ad66e4abdd903b0ed153c3cf71a6a8066e97f9/packages/pretty-format/src/plugins/DOMElement.ts.

Also had to increase the timeout for this test since it uses
wait/waitForElement several times.