Test behavior of being able to return multiple synteny features if there is a gene duplication, something that was not really done properly on the synteny_demo branch but which still was able to be displayed.
The way the model for fetching synteny data is that if there is a gene duplication in peach e.g. G->P1, G->P2 then two SimpleFeatures will be returned when fetching the region where G is from grape

This allows us to connect features by their "syntenyId" in a simple way.
The alternative is to always have an array of "syntenyIds" but this to
me complicates things, whereas a duplication is more of a special
situation and this behavior seems valid