Add marks for valid zoom levels to ZoomControls
Since zoom levels are constrained, this uses marks on the zoom slider to
show marks where those zoom level are and constrain the slider to those
values. We may move away from constrained zoom levels in the future, but
until then this better represents the actual state of the app.