Add ability to have multiple snackbar messages
The messages only show one at a time. If there is an existing message
when a new message is added, the  new message replaces the older one
(with animation). When the new message is closed, the older one
re-appears (with animation). All messages must be dismissed for the
snackbar to disappear. This allows the user to have time to review all
messages, even if many are added quickly in succession.