Zoom in sort test so read order doesn't matter
Since node < v11 uses an unstable sort, zoom in sort snapshot test so
that for that block, all features are the width of the block and don't
have any subfeatures. This means that as long as the read strand order
is sorted correctly, the order of individual reads don't matter.