Create plaintext GtfAdapter in plugins/gtf (#2389)
Note: No tabix version of GTF is available because it doesn't have parent features that can be redispatched on. No text indexing enabled by default also for similar reasons e.g. the lines aren't substantial enough to index, though if the hierarchies are reconstructed, it might be ok to restore